Update AT1-Lecture8.tex
This drawing is very abstract, it totally needs a few words explaining it. Again, the tutor used time as a third dimension of the drawing, which i cant.

The explanation goes something like this: as long as there is a square with at least one side fixed and one side not fixed, we can lift it. If you do this process row by row, you can totally lift the whole square. If you do not do it carefully, you reach a situation in which a square can't be lifted.