Create Lec9Pic2.tex
The picture i never did and with an explanation. Feel free to modify the explanation as to make it match your writing style. I feel it is necessary though. If you ever switch to the Figure environments, the caption should explain what we are seeing (O_x is the big blob, K_x is the sharp darker blob, etc).